fix bug loginModal
