+++ front-end/src/components/layout/layout.jsx
@@ -6,8 +6,13 @@ import Header from "../home/header/header";
 import Footer from "../home/footer/footer";
 import LoginModal from "../login/loginModal";
 import { useLoginModal } from "../../contexts/LoginModalContext";
+import { useAuth } from '../../contexts/AuthContext';
 const Layout = () => {
-    const {isLoginModalOpen, closeLoginModel} = useLoginModal();
+    const { isLoginModalOpen, openLoginModal, closeLoginModal } = useLoginModal();
+    const { login } = useAuth();
+    const handleLoginSuccess = () => {
+        login();
+    };
     return (
         <div>
             <header className="sticky top-0 z-50 bg-white shadow-md">
@@ -16,7 +21,9 @@ const Layout = () => {
             <ScrollRestoration />
             <Outlet />
             <Footer />
-            {isLoginModalOpen && <LoginModal onClose={closeLoginModel} />}
+            {isLoginModalOpen && <LoginModal 
+                onClose={closeLoginModal} 
+                onLoginSuccess={handleLoginSuccess}/>}
         </div>
     );
 };
